Don't show notification count after navigating to view.
This commit fixes the issue that users expect the notification count label to go away after they've navigated to the page (however, the notification count will re-appear if new notifications come in). 

One issue is that we don't adjust the notification count to the number of *new* notifications, it will always appear as the total number of notifications (even those already seen).